fix: newsletter signup form
